integrate abstractions regressor interface
