+++ openstef/model/regressors/regressor.py
@@ -5,11 +5,11 @@ from typing import Optional
 
 import numpy as np
 import pandas as pd
+from abc import abstractmethod
+from sklearn.base import BaseEstimator, RegressorMixin
 
-from openstef.model.regressors.regressor_interface import OpenstfRegressorInterface
 
-
-class OpenstfRegressor(OpenstfRegressorInterface):
+class OpenstfRegressor(BaseEstimator):
     def __init__(self):
         self.feature_importance_dataframe = None
         self.feature_importances_ = None
@@ -53,3 +53,46 @@ class OpenstfRegressor(OpenstfRegressorInterface):
 
         """
         return None
+
+    ## Define abstract methods required to be implemented by concrete models
+    @property
+    @abstractmethod
+    def feature_names(self):
+        """Retrieve the model input feature names
+
+        Returns:
+            The list of feature names
+        """
+        pass
+
+    @property
+    @abstractmethod
+    def can_predict_quantiles(self) -> bool:
+        """Can the model predict particular quantiles"""
+        pass
+
+    @abstractmethod
+    def predict(self, x: pd.DataFrame, **kwargs) -> np.array:
+        """Makes a prediction. Only available after the model has been trained
+        Args:
+            x (np.array): Feature matrix
+            kwargs: model-specific keywords
+
+        Returns:
+            (np.array): prediction
+        """
+        pass
+
+    @abstractmethod
+    def fit(self, x: np.array, y: np.array, **kwargs) -> RegressorMixin:
+        """Fits the regressor
+
+        Args:
+            x (np.array): Feature matrix
+            y (np.array): Labels
+            kwargs: model-specific keywords
+
+        Returns:
+            Fitted model
+        """
+        pass
